DeepBruce_AI/requirements.txt
flask
flask-cors
openai>=1.40.0
requests
beautifulsoup4
networkx
python-dotenv
rank-bm25
scikit-learn
